namepace update from CentralMail to BenefitsIntake (#19595)
* namepace update from CentralMail to BenefitsIntake

* updated remaining CentralMail items

* fixed linting

* removes rubocop disable from monitor spec

* removed rubocop disable
