+++ app/sidekiq/bgs/submit_form686c_job.rb
@@ -81,9 +81,11 @@ module BGS
 
     def self.send_backup_submission(vet_info, saved_claim_id, user_uuid)
       user = generate_user_struct(vet_info)
-      CentralMail::SubmitCentralForm686cJob.perform_async(saved_claim_id,
-                                                          KmsEncrypted::Box.new.encrypt(vet_info.to_json),
-                                                          KmsEncrypted::Box.new.encrypt(user.to_h.to_json))
+      Lighthouse::BenefitsIntake::SubmitCentralForm686cJob.perform_async(
+        saved_claim_id,
+        KmsEncrypted::Box.new.encrypt(vet_info.to_json),
+        KmsEncrypted::Box.new.encrypt(user.to_h.to_json)
+      )
       InProgressForm.destroy_by(form_id: FORM_ID, user_uuid:)
     rescue => e
       Rails.logger.warn('BGS::SubmitForm686cJob backup submission failed...',
